requirements.txt
amqp>=5.1.1
asgiref>=3.7.2
async-timeout>=4.0.2
attrs>=23.1.0
billiard>=4.1.0
boto3>=1.26.155
botocore>=1.29.155
celery>=5.3.0
click>=8.1.3
click-didyoumean>=0.3.0
click-plugins>=1.1.1
click-repl>=0.3.0
Django==4.2.14
django-cleanup>=8.0.0
django-cors-headers>=4.1.0
django-dbbackup>=4.0.2
django-rest-passwordreset>=1.3.0
django-storages>=1.13.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
drf-spectacular>=0.26.2
Faker>=18.10.1
flake8>=6.0.0
inflection>=0.5.1
jmespath>=1.0.1
jsonschema>=4.17.3
kombu>=5.3.1
mccabe>=0.7.0
packaging>=21.3
Pillow>=10.0.1
prompt-toolkit>=3.0.38
psycopg==3.2.1
pycodestyle>=2.10.0
pyflakes>=3.0.1
PyJWT>=2.7.0
pyparsing>=3.0.9
pyrsistent>=0.19.3
python-dateutil>=2.8.2
python-decouple>=3.8
pytz>=2023.3
PyYAML>=6.0
redis>=4.5.5
s3transfer>=0.6.1
six>=1.16.0
sqlparse>=0.4.4
tomli>=2.0.1
typing_extensions>=4.6.3
tzdata>=2023.3
uritemplate>=4.1.1
urllib3>=1.26.16
vine>=5.0.0
wcwidth>=0.2.6
